Move bloom into separate crate
