+++ crates/core/Cargo.toml
@@ -31,6 +31,7 @@ axum-macros = { workspace = true }
 base64 = { workspace = true }
 bincode = { workspace = true }
 bitvec = { workspace = true }
+bloom = { path = "../bloom" }
 bytecount = { workspace = true }
 bytemuck = { workspace = true }
 byteorder = { workspace = true }
